requirements.txt
alembic==0.9.1
appdirs==1.4.3
Babel==2.4.0
backlash==0.1.2
Beaker==1.8.1
crank==0.8.1
decorator==4.0.11
FormEncode==2.0.0a1
gearbox==0.1.1
Kajiki==0.6.1
Mako==1.0.6
MarkupSafe==1.0
nine==1.0.0
olefile==0.44
packaging==16.8
PasteDeploy==1.5.2
Pillow==4.1.0
preview-generator==0.1
Pygments==2.2.0
pyparsing==2.2.0
python-editor==1.0.3
python-magic==0.4.13
pytz==2017.2
repoze.lru==0.6
repoze.who==2.3
six==1.10.0
speaklater==1.3
sprox==0.10.2
SQLAlchemy==1.1.9
Tempita==0.5.2
tg.devtools==2.3.10
tgext.admin==0.7.4
tgext.crud==0.8.2
tgext.debugbar==0.3.0
transaction==2.1.2
TurboGears2==2.3.10
tw2.core==2.2.4
tw2.forms==2.2.5
WebHelpers2==2.0
WebOb==1.7.2
WebTest==1.4.3
zope.interface==4.3.3
zope.sqlalchemy==0.7.7
